+++ R/flextable-package.R
@@ -6,15 +6,13 @@
 #'
 #' The documentation of functions can be opened with command `help(package = "flextable")`.
 #'
-#' To learn more about flextable, start with the vignettes:
-#' `browseVignettes(package = "flextable")`.
-#'
 #' `flextable()` function is producing flexible tables where each cell
 #' can contain several chunks of text with their own set of formatting
 #' properties (bold, font color, etc.). Function [compose()] lets customise
 #' text of cells.
 #'
-#' @seealso <https://davidgohel.github.io/flextable/>, [flextable()]
+#' @seealso <https://davidgohel.github.io/flextable/>,
+#' <https://ardata-fr.github.io/flextable-book/>, [flextable()]
 #' @docType package
 #' @aliases flextable-package
 #' @name flextable-package